Update of continuity proof to be chain-specific
Continuity of functions is only relevant for chain subsets of the parameters. Since D is a ccpo, it cannot be assumed that there is a suprema for D. The same is true for [D -> D]. Because of that, the proof must work with:
1) F being a subset of [D -> D] such that F is a chain
2) C being a subset of C such that C is a chain
whenever a suprema is required. Since continuity only requires the preserval of supremas of chains, this assumption is sufficient when performing the proof.

Furthermore, it has come to my attantion that expressions such as
  f ( D ) = f d \in D | \forall f \in F
can be confusing. It is not clear, if the "| ..." part applies to only the right part of the equation or to the entire equation as a whole. In order to cope with this confusion I hhave established, that in this notation it only applies to the right. In order to have it apply to the entire equation, the equation itself must be wrapped in brackets
-> this is still issue for discussion. Will give it some thought till tomorrow.