Finished "attribute grammar"-like solution for exercise 5
For exercise 5, the "attribute grammar"-like solution has been finished, displaying the denotational semantics as functions that are assigned to the "statechange" attributes of the individual language elements. The functions are displayed using a lambda-notation.

Furthermore, I established that composite functions such as f(g(x)) are always displayed as (f ° g)(x). This is slightly more readable and especially helpful when for either f or g no function name exists, but only the function body. Exercise 6 was changed accordingly.